Properly handle namespaced resources
The proper field for a namespace on a resource is namespaceId, not
namepsace.

The id of a namespaced resource is namespace:name, which doesn't
match k8s auth for k8s rules where resourceName is set, because those
names are not prepended with the namespce, so strip the namespace from
the id before checking.